Add files via upload
Yeah, I was kinda bored so I've just made the program, I know its kinda messy, but still  I think its decent at least. I'm more intrested in your thoughts on it, so hope you'll like it. I'll probably get going with the rest of these projects in no time, so wish me luck I guess, bc It will probably take me a lot of time to learn new things and to make the old ones better. I'm really hoping that I will be able to make it and with that great speech out of the way, here it goes, I know its nothing for you but I'm pretty proud of the progress I've made to be honest.